refactor: 💡 sidebar, navbar, home and about
Refactor and make some files smaller
