+++ src/components/ui/sidebar/sidebar-context.tsx
@@ -1,0 +1,4 @@
+import * as React from 'react'
+import type { SidebarContext } from './sidebar-types'
+
+export const SidebarContextValue = React.createContext<SidebarContext | null>(null)
