+++ src/components/ui/sidebar/sidebar-hooks.tsx
@@ -1,0 +1,11 @@
+import { useContext } from 'react'
+import { SidebarContextValue } from './sidebar-context'
+
+export function useSidebar() {
+  const context = useContext(SidebarContextValue)
+  if (!context) {
+    throw new Error('useSidebar must be used within a SidebarProvider.')
+  }
+
+  return context
+}
