+++ src/data/about.ts
@@ -1,0 +1,83 @@
+export const educationTimeline = [
+  {
+    year: '2022',
+    title: 'App Development for iOS & Android',
+    organization: 'IT-Högskolan, Gothenburg',
+    description:
+      'Completed studies in mobile application development, focusing on iOS and Android platforms with modern development practices.',
+  },
+  {
+    year: '2022',
+    title: 'Internship - App Developer',
+    organization: 'Gardenize AB, Göteborg',
+    description:
+      'Second internship focusing on iOS development and Android features. Worked on native apps and implemented push notifications for garden management features.',
+  },
+  {
+    year: '2021',
+    title: 'Internship - System Developer',
+    organization: 'Find Sourcing, Gothenburg',
+    description:
+      'First internship creating applications from scratch using JavaScript and React Native, focusing on user-friendly tariff code solutions.',
+  },
+]
+
+export const experienceTimeline = [
+  {
+    year: '2022',
+    title: 'Software Developer',
+    organization: 'Capio AB',
+    description:
+      'Working as a junior software developer with the whole stack. Developed both front-end and back-end solutions using React, React Native, JavaScript, TypeScript, C#, and .NET.',
+  },
+]
+
+export const skills = [
+  {
+    category: 'Mobile Development',
+    technologies: ['React Native', 'iOS', 'Android', 'Swift', 'Java'],
+  },
+  {
+    category: 'Frontend',
+    technologies: ['React', 'Vue', 'TypeScript', 'JavaScript', 'HTML/CSS'],
+  },
+  {
+    category: 'Backend',
+    technologies: ['C#', '.NET', 'Firebase', 'MySQL', 'REST APIs'],
+  },
+  {
+    category: 'Tools & Methods',
+    technologies: [
+      'Xcode',
+      'Android Studio',
+      'Git',
+      'Objective-C',
+      'Database Functionality, AWS, CI/CD & Docker',
+    ],
+  },
+]
+
+export const personalStrengths = [
+  {
+    title: 'Full-Stack Development',
+    description:
+      'Experience working with both frontend and backend technologies, from mobile apps to web solutions, ensuring complete project delivery.',
+  },
+  {
+    title: 'Cross-Platform Solutions',
+    description:
+      'Skilled in developing applications for multiple platforms using React Native, ensuring consistent user experience across iOS and Android.',
+  },
+  {
+    title: 'Multilingual Communication',
+    description:
+      'Fluent in Swedish, English, and Serbo-Croatian, enabling effective communication in diverse international teams and projects.',
+  },
+  {
+    title: 'Healthcare Technology',
+    description:
+      'Specialized experience in healthcare technology through my work at Capio, understanding the unique requirements of medical applications.',
+  },
+]
+
+export const profileImage = '/uploads/07d45f1f-1633-4df8-92fa-a6306a59bbb5.png'
