+++ src/data/contact.ts
@@ -1,0 +1,9 @@
+export const contactInfo = {
+  email: 'romandivkovic@outlook.com',
+  phone: '+46 72 175 41 73',
+  location: 'Gothenburg, Sweden',
+  socialLinks: {
+    github: 'https://github.com',
+    linkedin: 'https://linkedin.com',
+  },
+}
