+++ src/hooks/use-about-data.ts
@@ -1,0 +1,8 @@
+import { educationTimeline, experienceTimeline } from '@/data/about'
+
+export const useAboutData = () => {
+  return {
+    educationTimeline,
+    experienceTimeline,
+  }
+}
