+++ src/hooks/use-scroll-detection.tsx
@@ -1,0 +1,16 @@
+import { useState, useEffect } from 'react'
+
+export const useScrollDetection = (threshold: number = 10) => {
+  const [isScrolled, setIsScrolled] = useState(false)
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setIsScrolled(window.scrollY > threshold)
+    }
+
+    window.addEventListener('scroll', handleScroll)
+    return () => window.removeEventListener('scroll', handleScroll)
+  }, [threshold])
+
+  return isScrolled
+}
