+++ src/lib/firebase.ts
@@ -2,13 +2,13 @@ import { initializeApp } from 'firebase/app'
 import { getFirestore } from 'firebase/firestore'
 
 const firebaseConfig = {
-  apiKey: 'AIzaSyA8JvGIocQjoE09vgK2pieCGJDFyWXYlGo',
-  authDomain: 'my-website-df82a.firebaseapp.com',
-  databaseURL: 'https://my-website-df82a-default-rtdb.europe-west1.firebasedatabase.app',
-  projectId: 'my-website-df82a',
-  storageBucket: 'my-website-df82a.appspot.com',
-  messagingSenderId: '423507224563',
-  appId: '1:423507224563:web:f5ff75952dee174d36d486',
+  apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
+  authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
+  databaseURL: import.meta.env.VITE_FIREBASE_DATABASE_URL,
+  projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
+  storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
+  appId: import.meta.env.VITE_FIREBASE_APP_ID,
 }
 
 const app = initializeApp(firebaseConfig)
