+++ src/pages/Contact.tsx
@@ -1,97 +1,12 @@
-import { useState, useEffect } from 'react'
-// Polyfill confetti for non-browser (test) environments
-const isTest = typeof process !== 'undefined' && process.env && process.env.NODE_ENV === 'test'
-let confetti: (...args: unknown[]) => void = () => {}
-if (!isTest && typeof window !== 'undefined') {
-  import('canvas-confetti').then((mod) => {
-    confetti = mod.default
-  })
-}
-import { useForm } from 'react-hook-form'
-import { zodResolver } from '@hookform/resolvers/zod'
-import * as z from 'zod'
-import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-import { Textarea } from '@/components/ui/textarea'
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form'
-import { toast } from '@/hooks/use-toast'
-import { Github, Linkedin, Mail, MapPin, Phone } from 'lucide-react'
+import { useEffect } from 'react'
 import Section from '@/components/Section'
-import emailjs from '@emailjs/browser'
-import { SERVICE_ID, TEMPLATE_ID, PUBLIC_KEY } from '../emailjs.config'
-
-// Define the form schema
-const formSchema = z.object({
-  name: z.string().min(2, { message: 'Name must be at least 2 characters.' }).max(100),
-  email: z.string().email({ message: 'Please enter a valid email address.' }),
-  message: z.string().min(10, { message: 'Message must be at least 10 characters.' }).max(1000),
-})
-
-type FormValues = z.infer<typeof formSchema>
+import { ContactForm } from '@/components/contact/ContactForm'
+import { ContactInfo } from '@/components/contact/ContactInfo'
 
 const Contact = () => {
-  const [isSubmitting, setIsSubmitting] = useState(false)
-  const [showSuccess, setShowSuccess] = useState(false)
-
   useEffect(() => {
     window.scrollTo(0, 0)
-    // Ingen init behövs för @emailjs/browser om du skickar publicKey i send
   }, [])
-
-  // Initialize form
-  const form = useForm<FormValues>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      name: '',
-      email: '',
-      message: '',
-    },
-  })
-
-  // Hantera formulärskick med EmailJS
-  const onSubmit = async (data: FormValues) => {
-    setIsSubmitting(true)
-    try {
-      await emailjs.send(
-        SERVICE_ID,
-        TEMPLATE_ID,
-        {
-          name: data.name,
-          message: data.message,
-          title: 'Contact Form',
-          email: data.email,
-        },
-        PUBLIC_KEY
-      )
-      toast({
-        title: 'Message sent successfully!',
-        description: "Thank you for reaching out. I'll get back to you soon.",
-      })
-      confetti({
-        particleCount: 80,
-        spread: 70,
-        origin: { y: 0.6 },
-      })
-      setShowSuccess(true)
-      setTimeout(() => setShowSuccess(false), 3000)
-      form.reset()
-    } catch (error) {
-      console.error(error)
-      toast({
-        title: 'Something went wrong!',
-        description: 'Could not send your message. Please try again later.',
-        variant: 'destructive',
-      })
-    }
-    setIsSubmitting(false)
-  }
 
   return (
     <div className="pt-16">
@@ -108,162 +23,8 @@ const Contact = () => {
             </p>
 
             <div className="grid grid-cols-1 lg:grid-cols-2 gap-12">
-              {/* Contact Form */}
-              <div className="bg-card shadow-sm rounded-lg p-6 md:p-8 animate-fade-in">
-                <h2 className="text-2xl font-bold mb-6">Send me a message</h2>
-
-                <Form {...form}>
-                  <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-                    <FormField
-                      control={form.control}
-                      name="name"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Name</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Your name" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name="email"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Email</FormLabel>
-                          <FormControl>
-                            <Input placeholder="your.email@example.com" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <FormField
-                      control={form.control}
-                      name="message"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Message</FormLabel>
-                          <FormControl>
-                            <Textarea
-                              placeholder="Your message here..."
-                              className="min-h-[150px]"
-                              {...field}
-                            />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-
-                    <Button
-                      type="submit"
-                      className="w-full flex items-center justify-center"
-                      disabled={isSubmitting}
-                    >
-                      {isSubmitting ? (
-                        <>
-                          <span className="mr-2">Sending</span>
-                          <span className="inline-block w-4 h-4 border-2 border-t-2 border-t-transparent border-primary rounded-full animate-spin"></span>
-                        </>
-                      ) : (
-                        'Send Message'
-                      )}
-                    </Button>
-                    {showSuccess && (
-                      <div className="mt-6 flex flex-col items-center animate-fade-in">
-                        <svg
-                          className="w-16 h-16 text-green-500 mb-2"
-                          fill="none"
-                          stroke="currentColor"
-                          strokeWidth="2"
-                          viewBox="0 0 24 24"
-                        >
-                          <path strokeLinecap="round" strokeLinejoin="round" d="M5 13l4 4L19 7" />
-                        </svg>
-                        <div className="text-lg font-semibold text-green-600">
-                          Message sent! Thank you 🙌
-                        </div>
-                      </div>
-                    )}
-                  </form>
-                </Form>
-              </div>
-
-              {/* Contact Info */}
-              <div className="space-y-8 animate-slide-in-right">
-                <h2 className="text-2xl font-bold mb-6">Contact Information</h2>
-
-                <div className="space-y-6">
-                  <div className="flex items-start">
-                    <div className="bg-muted rounded-full p-3 mr-4">
-                      <Mail className="h-5 w-5 text-secondary" />
-                    </div>
-                    <div>
-                      <h3 className="font-medium mb-1">Email</h3>
-                      <a
-                        href="mailto:romandivkovic@outlook.com"
-                        className="text-muted-foreground hover:text-secondary transition-colors"
-                      >
-                        romandivkovic@outlook.com
-                      </a>
-                    </div>
-                  </div>
-
-                  <div className="flex items-start">
-                    <div className="bg-muted rounded-full p-3 mr-4">
-                      <Phone className="h-5 w-5 text-secondary" />
-                    </div>
-                    <div>
-                      <h3 className="font-medium mb-1">Phone</h3>
-                      <a
-                        href="tel:+46721754173"
-                        className="text-muted-foreground hover:text-secondary transition-colors"
-                      >
-                        +46 72 175 41 73
-                      </a>
-                    </div>
-                  </div>
-
-                  <div className="flex items-start">
-                    <div className="bg-muted rounded-full p-3 mr-4">
-                      <MapPin className="h-5 w-5 text-secondary" />
-                    </div>
-                    <div>
-                      <h3 className="font-medium mb-1">Location</h3>
-                      <p className="text-muted-foreground">Gothenburg, Sweden</p>
-                    </div>
-                  </div>
-                </div>
-
-                <div className="pt-8 border-t border-muted">
-                  <h3 className="font-medium mb-4">Connect with me</h3>
-                  <div className="flex space-x-4">
-                    <a
-                      href="https://github.com"
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="bg-muted hover:bg-secondary/20 transition-colors rounded-full p-3"
-                      aria-label="GitHub"
-                    >
-                      <Github className="h-5 w-5" />
-                    </a>
-                    <a
-                      href="https://linkedin.com"
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="bg-muted hover:bg-secondary/20 transition-colors rounded-full p-3"
-                      aria-label="LinkedIn"
-                    >
-                      <Linkedin className="h-5 w-5" />
-                    </a>
-                  </div>
-                </div>
-              </div>
+              <ContactForm />
+              <ContactInfo />
             </div>
           </div>
         </div>
